requirements.txt
google-adk
mcp
httpx
anyio
google-cloud-storage
